refactor: turn index into ts file and export public types
