+++ packages/decap-cms-core/package.json
@@ -5,13 +5,12 @@
   "repository": "https://github.com/decaporg/decap-cms/tree/main/packages/decap-cms-core",
   "bugs": "https://github.com/decaporg/decap-cms/issues",
   "module": "dist/esm/index.js",
+  "types": "dist/esm/index.d.ts",
   "main": "dist/decap-cms-core.js",
   "files": [
     "src/",
-    "dist/",
-    "index.d.ts"
+    "dist/"
   ],
-  "types": "index.d.ts",
   "scripts": {
     "develop": "npm run build:esm -- --watch",
     "webpack": "node --max_old_space_size=4096 ../../node_modules/webpack/bin/webpack.js",
