+++ packages/decap-cms-core/src/index.ts
@@ -1,7 +1,10 @@
 import bootstrap from './bootstrap';
 import Registry from './lib/registry';
 
-export const DecapCmsCore = {
+import type { CMS } from './types';
+export * from './types';
+
+export const DecapCmsCore: CMS = {
   ...Registry,
   init: bootstrap,
 };
